Fail when nomad_server and nomad_client both set
and emit a helpful error message.

According to https://nomadproject.io/docs/agent/config.html, "Note that
it is strongly recommended not to operate a node as both client and
server".

Without this, someone can easily set `nomad_server` to true, not
realizing that `nomad_client` is set to true by default, and the result
of this is that the playbook runs successfully but no process is started
on the target node because it ends up with an upstart config that has no
script to run in it. This is a bit confusing.

Fixes: GH-5